+++ src/components/TimeSeriesPanel/TimeSeriesChartHeader.tsx
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-import React, { useRef, useState } from "react";
+import React, { RefObject, useRef, useState } from "react";
 import Box from "@mui/material/Box";
 import CircularProgress from "@mui/material/CircularProgress";
 import IconButton from "@mui/material/IconButton";
@@ -39,6 +39,7 @@ import FitScreenIcon from "@mui/icons-material/FitScreen";
 import IsoIcon from "@mui/icons-material/Iso";
 import ScatterPlotIcon from "@mui/icons-material/ScatterPlot";
 import ShowChartIcon from "@mui/icons-material/ShowChart";
+import CameraAltIcon from "@mui/icons-material/CameraAlt";
 
 import i18n from "@/i18n";
 import {
@@ -49,8 +50,10 @@ import {
 import { WithLocale } from "@/util/lang";
 import { makeStyles } from "@/util/styles";
 import { TimeSeriesChartType } from "@/states/controlState";
+import { MessageType } from "@/states/messageLogState";
 import TimeSeriesAddButton from "./TimeSeriesAddButton";
 import ValueRangeEditor from "./ValueRangeEditor";
+import { ExportOptions, exportElement } from "@/util/export";
 
 type ValueRange = [number, number];
 const SHOW_DEV_VALUE = "stddev";
@@ -86,6 +89,7 @@ const styles = makeStyles({
     paddingRight: theme.spacing(1),
   }),
 });
+const EXPORT_WIDTH = 2000; // in pixels
 
 interface TimeSeriesChartHeaderProps extends WithLocale {
   timeSeriesGroup: TimeSeriesGroup;
@@ -109,6 +113,8 @@ interface TimeSeriesChartHeaderProps extends WithLocale {
   setStdevBars: (showStdDev: boolean) => void;
   valueRange: ValueRange | undefined;
   setValueRange: (fixedValueRange: ValueRange | undefined) => void;
+  chartElement: RefObject<HTMLDivElement>;
+  postMessage: (messageType: MessageType, messageText: string | Error) => void;
 }
 
 export default function TimeSeriesChartHeader({
@@ -130,13 +136,14 @@ export default function TimeSeriesChartHeader({
   setStdevBars,
   valueRange,
   setValueRange,
+  chartElement,
+  postMessage,
 }: TimeSeriesChartHeaderProps) {
   const valueRangeEl = useRef<HTMLButtonElement | null>(null);
   const [valueRangeEditorOpen, setValueRangeEditorOpen] = useState(false);
   const timeSeriesText = i18n.get("Time-Series");
   const unitsText = timeSeriesGroup.variableUnits || i18n.get("unknown units");
   const chartTitle = `${timeSeriesText} (${unitsText})`;
-
   const handleToggleValueRangeEditor = () => {
     setValueRangeEditorOpen(!valueRangeEditorOpen);
   };
@@ -161,6 +168,23 @@ export default function TimeSeriesChartHeader({
       values.length === 1 ? (values[0] as TimeSeriesChartType) : chartType,
     );
     setStdevBars(showStdDevNew);
+  };
+
+  const handleExportSuccess = () => {
+    postMessage("success", i18n.get("Snapshot copied to clipboard"));
+  };
+
+  const handleExportError = (error: unknown) => {
+    const message = "Error copying snapshot to clipboard";
+    console.error(message + ":", error);
+    postMessage("error", i18n.get(message));
+  };
+
+  const exportOptions: ExportOptions = {
+    format: "png",
+    width: EXPORT_WIDTH,
+    handleSuccess: handleExportSuccess,
+    handleError: handleExportError,
   };
 
   return (
@@ -256,6 +280,20 @@ export default function TimeSeriesChartHeader({
             </ToggleButton>
           </Tooltip>
         </ToggleButtonGroup>
+
+        <Tooltip arrow title={i18n.get("Copy snapshot of chart to clipboard")}>
+          <IconButton
+            key={"exportButton"}
+            sx={styles.actionButton}
+            onClick={() =>
+              chartElement.current &&
+              exportElement(chartElement.current, exportOptions)
+            }
+            size="small"
+          >
+            <CameraAltIcon fontSize={"inherit"}></CameraAltIcon>
+          </IconButton>
+        </Tooltip>
 
         <TimeSeriesAddButton
           sx={styles.actionButton}